+++ src/ServiceOptions.php
@@ -1,0 +1,39 @@
+<?php
+
+/**
+ * This file is part of the Phalcon Talon.
+ *
+ * (c) Phalcon Team <team@phalcon.io>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Phalcon\Talon;
+
+final class ServiceOptions
+{
+    /**
+     * @param array<string, mixed> $options
+     */
+    public function __construct(
+        private string $name,
+        private array $options,
+    ) {
+    }
+
+    public function getName(): string
+    {
+        return $this->name;
+    }
+
+    /**
+     * @return array<string, mixed>
+     */
+    public function getOptions(): array
+    {
+        return $this->options;
+    }
+}